requirements.txt
openai>=1.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
duckduckgo-search>=6.0.0
pytest>=7.4.0
llama-cpp-python>=0.2.0
